% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSMOTE.R
\name{RSMOTE}
\alias{RSMOTE}
\title{RSMOTE, an over-sampler for imbalanced data}
\usage{
RSMOTE(Data, IR = 0.8, k = 6)
}
\arguments{
\item{Data}{A data frame or matrix of numeric-attributed dataset,with the first row indicating the two classes with 0 and 1.}

\item{IR}{The proportion of minority and majority samples.}

\item{k}{The number of nearest neighbors during sampling process.}
}
\value{
The synthetic samples as a matrix.
}
\description{
Generate synthetic samples for imbalanced data with two classes,using RSMOTE algorithm. Synthetic samples are generated from minority class.
}
\details{
Data: each column represents one sample, the first element represent its class;
the number of synthetic samples = IR * number of majority samples - number of minority samples.
}
\examples{
library(RSMOTE)
mat1 = matrix(nrow=3, ncol=1000)
mat1[1,] = 1
mat1[2,] = c(rnorm(1000,1,0.5))
mat1[3,] = c(rnorm(1000,0.6,0.6))
mat2 = matrix(nrow=3, ncol=100)
mat2[1,] = 0
mat2[2,] = c(rnorm(100,2.5,0.7))
mat2[3,] = c(rnorm(100,2.5,0.5))
mat = cbind(mat1,mat2)
synthetic = RSMOTE(mat, IR = 0.9, k = 5)
library(ggplot2)
ggplot()+
geom_point(aes(mat2[2,],mat2[3,]),col='blue')+
geom_point(aes(mat1[2,],mat1[3,]),col='red')+
geom_point(aes(synthetic[1,], synthetic[2,]), col='green')
}
